# [Rust Concurrent] Use flattened list