Slower...

Main: `| 12.47 ms  | 174.48 ms  | 1.52 s `

This PR:
```
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 15.084941ms
        Processing time (w/o IO): 14.97014ms
        Processing time (w/o IO): 15.523745ms
        Processing time (w/o IO): 15.255743ms
        Processing time (w/o IO): 15.137541ms
        Processing time (w/o IO): 14.96174ms
        Processing time (w/o IO): 15.520045ms
        Processing time (w/o IO): 15.031641ms
        Processing time (w/o IO): 15.103841ms
        Processing time (w/o IO): 15.199042ms
        Processing time (w/o IO): 14.96114ms
        Processing time (w/o IO): 15.068441ms
        Processing time (w/o IO): 14.97814ms
          Time (mean ± σ):      31.7 ms ±   0.3 ms    [User: 35.5 ms, System: 10.9 ms]
          Range (min … max):    31.0 ms …  32.0 ms    10 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 222.285776ms
        Processing time (w/o IO): 221.972472ms
        Processing time (w/o IO): 223.469286ms
          Time (mean ± σ):     281.3 ms ±   1.6 ms    [User: 476.9 ms, System: 23.4 ms]
          Range (min … max):   280.2 ms … 282.5 ms    2 runs

Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 1.987924765s
        Processing time (w/o IO): 1.977579869s
        Processing time (w/o IO): 1.978538079s
          Time (mean ± σ):      2.164 s ±  0.004 s    [User: 4.045 s, System: 0.084 s]
          Range (min … max):    2.162 s …  2.167 s    2 runs

```