@jinyus 2023-11-17
2nd run:
```
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 18.833407ms
        Processing time (w/o IO): 15.629572ms
        Processing time (w/o IO): 15.138167ms
        Processing time (w/o IO): 15.073367ms
        Processing time (w/o IO): 15.022566ms
        Processing time (w/o IO): 15.46377ms
        Processing time (w/o IO): 15.014166ms
        Processing time (w/o IO): 15.038366ms
        Processing time (w/o IO): 15.127767ms
        Processing time (w/o IO): 15.178668ms
        Processing time (w/o IO): 15.016566ms
        Processing time (w/o IO): 15.072366ms
        Processing time (w/o IO): 14.982666ms
          Time (mean ± σ):      31.3 ms ±   0.4 ms    [User: 35.8 ms, System: 10.2 ms]
          Range (min … max):    30.8 ms …  32.0 ms    10 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 227.279607ms
        Processing time (w/o IO): 225.479088ms
        Processing time (w/o IO): 224.646979ms
          Time (mean ± σ):     284.5 ms ±   0.5 ms    [User: 481.0 ms, System: 23.4 ms]
          Range (min … max):   284.2 ms … 284.9 ms    2 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 1.977287114s
        Processing time (w/o IO): 1.982330787s
        Processing time (w/o IO): 1.97996685s
          Time (mean ± σ):      2.174 s ±  0.003 s    [User: 4.074 s, System: 0.067 s]
          Range (min … max):    2.171 s …  2.176 s    2 runs
```