@jinyus 2023-11-20
Results from new VM

Main: `4.71 ms | 60.90 ms | 491.75 ms `

This PR:
```
Rust Concurrent:
        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 6.120289ms
        Processing time (w/o IO): 6.299317ms
        Processing time (w/o IO): 6.283657ms
        Processing time (w/o IO): 6.061148ms
        Processing time (w/o IO): 6.059789ms
        Processing time (w/o IO): 6.036409ms
        Processing time (w/o IO): 6.015789ms
        Processing time (w/o IO): 6.041168ms
        Processing time (w/o IO): 6.058318ms
        Processing time (w/o IO): 6.165858ms
        Processing time (w/o IO): 6.055739ms
        Processing time (w/o IO): 6.219638ms
        Processing time (w/o IO): 6.048319ms
          Time (mean ± σ):      18.6 ms ±   2.4 ms    [User: 28.4 ms, System: 5.8 ms]
          Range (min … max):    16.7 ms …  22.4 ms    10 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 86.870743ms
        Processing time (w/o IO): 87.188151ms
        Processing time (w/o IO): 87.607489ms
          Time (mean ± σ):     160.6 ms ±   4.8 ms    [User: 363.9 ms, System: 21.8 ms]
          Range (min … max):   157.2 ms … 163.9 ms    2 runs
         
Rust Concurrent:

        Benchmark 1: ./target/release/rust_rayon
        Processing time (w/o IO): 763.018583ms
        Processing time (w/o IO): 790.846529ms
        Processing time (w/o IO): 768.049986ms
          Time (mean ± σ):      1.038 s ±  0.018 s    [User: 3.149 s, System: 0.050 s]
          Range (min … max):    1.026 s …  1.050 s    2 runs
```